+++ datatree/tests/test_etreerender.py
@@ -13,7 +13,7 @@ class test_ETreeRenderer(unittest.TestCase):
     def test_single_level(self):
         tree = Tree()
         root = tree.a('Here', href='url', title='A Title')
-        actual = root.render('etree')
+        actual = root.render('etree', as_root=True)
         result = e.fromstring(actual)
 
         self.assertEqual(result.tag, 'a')
@@ -28,7 +28,7 @@ class test_ETreeRenderer(unittest.TestCase):
             two.third('Three', level=3)
         root.fourth('four', level=2)
 
-        actual = root.render('etree')
+        actual = root.render('etree', as_root=True)
 
         # Parse the xml and break it into pieces to assert.
         result = e.fromstring(actual)